Add Register service and controller with changes in the repositorys

+++ src/main/java/com/litografiaartesplanchas/orderservice/repository/OrdersRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.litografiaartesplanchas.orderservice.model.Orders;
 
 @Repository
-public interface OrdersRepository extends JpaRepository<Orders, Long> {
+public interface OrdersRepository extends JpaRepository<Orders, Integer> {
 
 }
